+++ metadata.rb
@@ -1,14 +1,15 @@
 name             'app-nginx'
 maintainer       'Earth U'
-maintainer_email 'iskitingbords @ gmail.com'
+maintainer_email 'iskitingbords@gmail.com'
 license          'Apache License'
 description      'Just a wrapper for setting up Nginx webserver'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 source_url       'https://github.com/nollieheel/app-nginx'
 issues_url       'https://github.com/nollieheel/app-nginx/issues'
-version          '1.0.0'
+version          '2.0.0'
 
-depends 'nginx', '~> 8.0.0'
-depends 'openssl', '~> 8.1.2'
+depends 'nginx', '~> 9.0'
+depends 'openssl', '~> 8.5'
 
 supports 'ubuntu', '14.04'
+supports 'ubuntu', '16.04'