+++ recipes/configure.rb
@@ -1,5 +1,5 @@
 #
-# Author:: Earth U (<iskitingbords @ gmail.com>)
+# Author:: Earth U (<iskitingbords@gmail.com>)
 # Cookbook Name:: app-nginx
 # Recipe:: configure
 #
@@ -20,7 +20,7 @@
 
 include_recipe 'openssl::upgrade'
 
-directory node['app-nginx']['priv_dir']
+directory(node['app-nginx']['priv_dir']) { recursive true }
 
 if node['app-nginx']['dhparam_path']
   openssl_dhparam node['app-nginx']['dhparam_path'] do
@@ -32,21 +32,14 @@ else
     action :delete
   end
 end
-#
-#node['app-nginx']['sites'].each do |site|
-#  site_main_name = site[:server_name].split(' ')[0]
-#  site_cb = site[:cookbook] || 'app-nginx'
-#  site_source = site[:source] || 'nginx_site.conf.erb'
-##
-#  template "#{node['app-nginx']['conf_dir']}/#{site_main_name}" do
-#    cookbook  site_cb
-#    source    site_source
-#    mode      0644
-#    notifies  :restart, 'service[nginx]', :delayed
-#    variables site
-#  end
-#
-#  nginx_site site_main_name do
-#    enable true
-#  end
-#end
+
+( node['app-nginx']['basic_auth'] || [] ).each do |e|
+  template e[:path] do
+    source    'htpasswd.erb'
+    mode      0644
+    sensitive true
+    variables(
+      :creds => e[:creds]
+    )
+  end
+end